Gwanghwamun_suggestion/requirements.txt
astroid==2.3.3
beautifulsoup4==4.6.0
boto==2.49.0
boto3==1.13.7
botocore==1.16.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
docutils==0.15.2
et-xmlfile==1.0.1
gensim==3.8.3
get==2019.4.13
idna==2.9
isort==4.3.21
jdcal==1.4.1
jmespath==0.9.5
joblib==0.14.1
JPype1==0.7.4
kiwisolver==1.2.0
konlpy==0.5.2
lazy-object-proxy==1.4.3
lxml==4.6.2
matplotlib==3.2.1
mccabe==0.6.1
nltk==3.5
numpy==1.18.3
oauthlib==3.1.0
openpyxl==3.0.3
pandas==1.0.3
Pillow==8.1.1
post==2019.4.13
public==2019.4.13
pylint==2.5.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
query-string==2019.4.13
regex==2020.4.4
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
six==1.14.0
smart-open==2.0.0
soupsieve==2.0
tqdm==4.45.0
tweepy==3.8.0
urllib3==1.25.9
wincertstore==0.2
wordcloud==1.7.0
wrapt==1.12.1
xlrd==1.2.0
